# apps/wiki/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plunk API",
    "description": "Open-source email platform API for transactional emails, campaigns, and marketing automation",
    "version": "1.0.0",
    "contact": {
      "name": "Plunk Support",
      "url": "https://useplunk.com"
    }
  },
  "servers": [
    {
      "url": "https://api.useplunk.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key authentication. **Secret keys (sk_*)** are required for all endpoints except `/v1/track`. **Public keys (pk_*)** only work with the `/v1/track` endpoint for client-side event tracking. The project is automatically derived from the key."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "time": {
            "type": "integer"
          }
        }
      }
    },
...
  },
  "paths": {
    "/v1/send": {
      "post": {
        "tags": ["Public API"],
        "summary": "Send transactional email",
        "description": "Send a transactional email via the public API. Automatically creates/updates contacts.",
        "operationId": "sendEmail",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to"],
                "properties": {
                  "to": {
                    "oneOf": [
                      {
                        "type": "string",
                        "format": "email"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    ],
                    "description": "Recipient email(s)"
                  },
                  "subject": {
                    "type": "string",
                    "description": "Email subject (required if no template)"
                  },
                  "body": {
                    "type": "string",
                    "description": "Email body HTML (required if no template)"
                  },
                  "template": {
                    "type": "string",
                    "description": "Template identifier to use"
                  },
                  "from": {
                    "type": "string",
                    "format": "email",
                    "description": "Custom from address (requires verified domain)"
                  },
                  "subscribed": {
                    "type": "boolean",
                    "description": "Whether recipient is subscribed to marketing emails. Defaults to false for transactional emails.",
                    "default": false
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Template variables"
                  },
                  "headers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Custom email headers"
                  },
                  "reply": {
                    "type": "string",
                    "format": "email",
                    "description": "Reply-to address"
                  },
                  "attachments": {
                    "type": "array",
                    "description": "Email attachments (max 10 attachments, 10MB total)",
                    "maxItems": 10,
                    "items": {
                      "type": "object",
                      "required": ["filename", "content", "contentType"],
                      "properties": {
                        "filename": {
                          "type": "string",
                          "maxLength": 255,
                          "description": "Attachment filename"
                        },
                        "content": {
                          "type": "string",
                          "description": "Base64 encoded file content"
                        },
                        "contentType": {
                          "type": "string",
                          "maxLength": 255,
                          "description": "MIME type (e.g., application/pdf, image/png)"
                        }
                      }
                    }
                  }
                }
              },
              "examples": {
                "simple": {
                  "summary": "Simple transactional email",
                  "value": {
                    "to": "user@example.com",
                    "subject": "Password Reset Request",
                    "body": "<h1>Reset Your Password</h1><p>Click the link to reset: {{resetLink}}</p>",
                    "data": {
                      "resetLink": "https://example.com/reset/abc123"
                    }
                  }
                },
                "withTemplate": {
                  "summary": "Using template",
                  "value": {
                    "to": "user@example.com",
                    "template": "welcome-email",
                    "data": {
                      "firstName": "John",
                      "lastName": "Doe"
                    }
                  }
                },
                "marketingEmail": {
                  "summary": "Marketing email (set subscribed: true)",
                  "value": {
                    "to": "user@example.com",
                    "subject": "Weekly Newsletter",
                    "body": "<h1>This Week's Updates</h1>",
                    "subscribed": true
                  }
                },
                "withAttachment": {
                  "summary": "Email with PDF attachment",
                  "value": {
                    "to": "user@example.com",
                    "subject": "Your Invoice",
                    "body": "<h1>Invoice Attached</h1><p>Please find your invoice attached.</p>",
                    "attachments": [
                      {
                        "filename": "invoice.pdf",
                        "content": "JVBERi0xLjQKJeLjz9MKMSAwIG9iago8PC9UeXBlL...",
                        "contentType": "application/pdf"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "emails": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "to": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/track": {
      "post": {
        "tags": ["Public API"],
        "summary": "Track event",
        "description": "Track an event for a contact. Automatically creates/updates the contact.",
        "operationId": "trackEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "event"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email"
                  },
                  "event": {
                    "type": "string",
                    "description": "Event name"
                  },
                  "subscribed": {
                    "type": "boolean",
                    "description": "Whether contact is subscribed to marketing emails. Defaults to true for event tracking.",
                    "default": true
                  },
                  "data": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Event data and contact metadata"
                  }
                }
              },
              "example": {
                "email": "user@example.com",
                "event": "purchase",
                "data": {
                  "product": "Premium Plan",
                  "amount": 99.0
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "contact": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            }
                          }
                        },
                        "event": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "event": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}